requirements.txt
clip==1.0
Flask==2.1.0
pandas==1.3.5
Pillow==9.3.0
torch==1.13.0
torch_neuron==1.0.1522.0
tqdm==4.64.1
